Updates   * Moved session list header to own component   * Added it as section header for each section
