+++ src/components/sessions.js
@@ -1,190 +1,15 @@
-import React, { useEffect, useCallback, useMemo } from 'react'
-import { useTheme, useDimensions, useStylesCallback } from '@keg-hub/re-theme'
-import { View, ItemHeader, Button, ScrollView } from '@keg-hub/keg-components'
+import React, { useEffect, useMemo } from 'react'
+import { useTheme } from '@keg-hub/re-theme'
+import { View } from '@keg-hub/keg-components'
 import { ModalManager } from 'SVComponents/modals/modalManager'
 import { mapSessionInterface } from 'SVActions/session/mapSessionInterface'
-import {
-  applySessionFilters,
-  clearSelectedFilters,
-} from 'SVActions/session/filters'
-import { incrementDay, decrementDay } from 'SVActions/session/dates'
-import { EmptyDayMessage } from 'SVComponents/grid/emptyDayMessage'
+import { applySessionFilters } from 'SVActions/session/filters'
 import { useStoreItems } from 'SVHooks/store/useStoreItems'
-import { useAgenda } from 'SVHooks/models/useAgenda'
-import { DayToggle } from 'SVComponents/dates/dayToggle'
 import { noOp } from 'SVUtils/helpers/method/noop'
-import { get } from '@keg-hub/jsutils'
-import { EVFIcons } from 'SVIcons'
 import { Values } from 'SVConstants'
-import { useCreateModal } from 'SVHooks/modal'
 import { SessionList } from 'SVComponents/sessionList'
 import { useBookingRequestEvent } from 'SVHooks/booking/useBookingRequestEvent'
 const { CATEGORIES, SUB_CATEGORIES } = Values
-
-/**
- * FilterButton
- * Renders either an Icon or a text button based on current screen dimension
- * @param {object} props
- * @param {object} props.styles
- * @param {Function} props.onClick
- * @param {boolean} props.showIcon
- */
-const FilterButton = ({ onClick, styles, showIcon }) => {
-  return showIcon ? (
-    <View
-      className={'ef-sessions-filter-button'}
-      style={styles?.filterIcon?.main}
-    >
-      <EVFIcons.Filter
-        className={'ef-sessions-filter-button'}
-        style={styles?.filterIcon?.icon}
-        onPress={onClick}
-        color={styles?.filterIcon?.icon?.color}
-      />
-    </View>
-  ) : (
-    <Button
-      className={'ef-sessions-filter-button'}
-      themePath='button.text.default'
-      styles={styles?.filterButton}
-      onClick={onClick}
-      content={'Filter'}
-    />
-  )
-}
-
-/**
- * Builds the styles for the header right component
- * @param {Object} theme - Global Theme object
- * @param {Object} custom - contains {styles, smallWidth}
- *
- * @returns {Object}
- */
-const buildStylesHeaderRight = (theme, custom) => {
-  return theme.get(
-    custom.styles,
-    custom.smallWidth && {
-      main: {
-        paddingRight: 0,
-      },
-    }
-  )
-}
-
-// TODO - Move this to the sessionList component
-// Add it as each section header
-/**
- * Component that will hold the day toggle and filter button
- * @param {object} props
- * @param {object} props.styles - styles obj
- * @param {Array<import('SVModels/label').Label>} props.labels - session labels
- * @param {Function} props.onDayChange - function for handling day changes in the day toggle
- */
-const SessionsHeader = ({ styles, onDayChange, labels }) => {
-  const {
-    agendaLength,
-    currentAgendaDay = {},
-    currentDayNumber,
-    isLatestDay,
-    isFirstDay,
-  } = useAgenda()
-
-  const increment = useCallback(() => incrementDay(onDayChange), [onDayChange])
-  const decrement = useCallback(() => decrementDay(onDayChange), [onDayChange])
-  const headerStyles = styles.content?.header
-  const displayFilterModal = useCreateModal(Values.MODAL_TYPES.FILTER, {
-    labels,
-  })
-
-  return (
-    <ItemHeader
-      styles={headerStyles}
-      CenterComponent={
-        <DayToggle
-          date={get(currentAgendaDay, 'date')}
-          dayNumber={currentDayNumber}
-          disableDecrement={isFirstDay}
-          disableIncrement={isLatestDay || !agendaLength}
-          onDecrement={decrement}
-          onIncrement={increment}
-        />
-      }
-      RightComponent={
-        <ItemHeaderRight
-          styles={headerStyles?.content?.right?.content}
-          onClick={displayFilterModal}
-        />
-      }
-    />
-  )
-}
-
-/**
- * ItemHeaderRight
- * displays the filter btn and clear btn
- *    hides the clear btn when dimension is small enough
- * @param {object} props
- * @param {object} props.styles - theme path: header.content.right.content
- */
-const ItemHeaderRight = ({ styles, onClick }) => {
-  const dim = useDimensions()
-  const activeFilters = useStoreItems(
-    `${CATEGORIES.FILTERS}.${SUB_CATEGORIES.ACTIVE_FILTERS}`
-  )
-
-  const smallWidth = dim.width <= 720
-  const showClearButton = dim.width > 520 && Boolean(activeFilters?.length)
-
-  const customStyles = useMemo(
-    () => ({
-      smallWidth,
-      styles,
-    }),
-    [ styles, smallWidth ]
-  )
-
-  const mainStyle = useStylesCallback(
-    buildStylesHeaderRight,
-    [ styles, smallWidth ],
-    customStyles
-  )
-  const clearActiveFilters = useCallback(() => {
-    clearSelectedFilters()
-    applySessionFilters()
-  }, [ applySessionFilters, clearSelectedFilters ])
-
-  return (
-    <View style={mainStyle?.main}>
-      <FilterButton
-        styles={mainStyle}
-        onClick={onClick}
-        showIcon={smallWidth}
-      />
-      { showClearButton && (
-        <Button
-          themePath='button.text.default'
-          styles={mainStyle?.clearAll}
-          content={'Clear'}
-          onClick={clearActiveFilters}
-        />
-      ) }
-    </View>
-  )
-}
-
-/**
- * Sets up the container for a group of sessions on a specific day
- * @param {object} props
- * @param {Array<import('SVModels/label').Label>} props.labels - session labels
- * @param {Array} props.daySessions - group of sessions by block. see buildHourSessionsMap helper
- * @param {boolean} props.enableFreeLabel - whether to display 'FREE' on session with no pricing or not
- * @param {boolean} props.militaryTime - whether to display time in 12 hr or 24 hr format
- * @returns {Component}
- */
-const AgendaSessions = React.memo(
-  ({ labels, daySessions, enableFreeLabel, militaryTime }) => {
-    if (!daySessions?.length) return <EmptyDayMessage />
-
 
 /**
  * SessionComponent
@@ -231,11 +56,6 @@ export const Sessions = props => {
       className={'ef-sessions-background'}
       style={sessionsStyles.main}
     >
-      <SessionsHeader
-        labels={labels}
-        styles={sessionsStyles}
-        onDayChange={onDayChange}
-      />
       <SessionList
         labels={labels}
         onDayChange={onDayChange}
