+++ src/components/sessionsHeader/index.js
@@ -1,0 +1,1 @@
+export * from './sessionsHeader'
